applications/product-service/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.19.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
flask-limiter==3.5.0